requirements.txt
django
django-extensions
djangorestframework
django-filter
markdown
ipdb
requests